Improve encoding dialog
This fixes various bugs and confusing behaviors of the encoding dialog,
and addresses #28, by always ending up proposing something to the user,
with the help of `g_utf8_make_valid()`.

Depending on the encoding possibilities that have been found, unused
widgets are hidden and the dialog header is updated consequently.

Closes #28, see !33 for more details.