Improve save-related actions
By fixing #67, this tries to clarify a bit the way the save-related
actions work, by connecting in particular some signals so that actions
and widgets are properly updated on file status change.

The static "Revert" action also becomes a dynamic "Reload/Revert"
action, which changes with the file modification status.

Fixes #67, see !39 for more details.