Improve tooltips management
This changes the way tooltips were handled in !25, by setting them
directly in `mousepad/resources/gtk/menus.ui` (except for the toolbar,
which is not based on the use of `GMenu`).

All the work to set tooltips at the window level is now done in the
single function `mousepad_window_menu_set_tooltips()`, which is called
only once to go through the whole menubar, and then only locally to
update submenus as needed.

See !38 for more details.