Add keybindings to show the menubar temporarily when hidden
This reproduces a common behavior for the menubar nowadays, typically
that of Firefox. If the menu bar is hidden:
* When the Alt key is released, show the menubar temporarily (i.e.
  settings are not affected);
* When the Alt or Esc keys are then pressed, hide the menubar;
* Hide also the menubar after some user actions (mouse buttons
  pressed/released, scroll, focus out);
* When a menubar mnemonic key is activated (e.g. Alt+f) show the
  menubar and popup the corresponding submenu (then hide the menubar
  and its submenu all at once).

By the way, this clarifies the way the menubar/toolbar/statusbar are
shown/hidden in window/fullscreen modes, by managing this in only one
common function `mousepad_window_update_bar_visibility()`.

Closes #47, see !42 for more details.