Add a GMenuModel-GtkToolbar binding
This last commit of the "GMenuModel-* binding" series adds a
`mousepad_window_toolbar_new_from_model()` function, in echo to the
`gtk_menu_bar_new_from_model()` function.

The goal is to build the same kind of binding, using again the
"items-changed" GMenuModel signal, and taking advantage of the use of
shared menu items introduced in the previous commit.

See !50 for more details.