Add a GMenuModel-GMenuModel binding
This GMenuModel-GMenuModel binding not only involves the use of the
"items-changed" signal, but also a trick in the way to write the
GtkBuilder menu files (`mousepad/resources/gtk/*.ui`).

The goal is to remove duplicates in these files, by defining a menu
item only once (with all its attributes: label, icon, tooltip, action,
accel, etc.), and then reusing it in any menu where it should appear,
using a syntax close to the expected but missing
`<item id="my-already-defined-item"/>`.

Downstream, this simplifies the synchronization of the different
widgets that use these shared menu items, and makes the overall
code easier to maintain.

See !50 for more details.