Switch to asynchronous search
This switches to asynchronous search, so as not to block the UI when
searching in big files and/or with regex search enabled. This is almost
required when the search mode is "find as you type", which is the case
for Mousepad.

This also improves the search widgets behavior when documents change
(from the buffer or tab), and the consistency between search widgets
when used together, within the limits of what is possible.

Codewise, this makes increased use of signals to connect the different
widgets involved in the search, from the search bar and replace dialog
to the documents, through the window and back. Some precautions are
also taken to avoid prohibitive computation times in some situations,
because of a GtkSourceView issue.

See !58 for more details.