Reset an idle function to default priority
This partially reverts 27c5ad98485c8924a972f8958deff160333d42a9, by
keeping only the existence test to not run a MousepadView function when
closing.

This is sufficient and in any case necessary to avoid side effects
described in !54, while the default idle priority has been more
widely tested since added in !13 to fix #26.

Furthermore, some tests when the search is asynchronous (see !58)
show that this default priority may be required to not reactivate #26.