Give more choice for whitespace display
This adds the ability to display only the white spaces at the beginning
of the line, the end of the line or between words (or any combination
of these three possibilities), accessible only from GSettings.

To do this, a property is added to the application that is synchronized
with GSettings on the one hand, and with the different views on the
other hand.

Closes #44, see !46 for more details.