Prevent regex search from running in the background
If regex search is enabled at startup, a useless and expensive search
can run in the background, when the file is loaded in the buffer. To
prevent this, the "regex-enabled" setting isn't bound at startup as
others, but bound/unbound with the "highlight" property when the search
widgets are shown/hidden.

This is a complement to the workaround already implemented in !58 to
avoid prohibitive computation times in some situations.