Add license text, remove obsolete code
This code that would extend the editing actions to the search bar (one
could also think of the replace dialog) was already no longer working
in Mousepad 0.4.2. It's not easy to make it work, because GtkEntry
doesn't have a good selection behavior, and it could also create focus
conflicts between widgets within the window, for a very questionable
gain in terms of functionality. So it's better to remove it.