Filter document signals on the active document
Only the active document is of interest here. It worked without this
filter until then because it was actually applied later, in the
functions called from the signal handlers. But it's neither secure nor
optimized.

This also fixes some wrong (or approximate) types for signal handlers
parameters, and enables the reload action when the document filename is
set for the first time (pointed out in
https://gitlab.xfce.org/apps/mousepad/-/merge_requests/64#note_20395).