Prevent too late accesses to the buffer
Now that the documents are finalized and no longer only destroyed when
the tabs are closed (see previous commit
585d7a43bc4fc9f573922831515be9dcaf33ad23), it turns out that the search
context can make too late accesses to the buffer, if a search is
launched when closing (typically following a tab switch) or a buffer
scanning is in progress.

As the search context does not keep a reference on the buffer until it
finishes its operations, this commit fills this gap by launching a last
trivial search when the document is finalized, which cancels any search
or buffer scanning in progress, and finally triggers the release of the
buffer and the search context.