Allow to choose encoding in the "Open" and "Save As" dialogs
In both cases, the combo box contains a shorten list with default,
system and current encodings, UTF-8 and ISO-8859-15 if not already
listed, and two actionable entries to show all encodings and open the
file in the encoding dialog.

For the latter action, in the case of the "Open" dialog, the first
selected file is opened in the encoding dialog. In the case of the
"Save As" dialog, the current buffer is saved in a temporary file to
perform the encoding tests.

Closes #42, see !69 for more details.