Display the current encoding in the status bar
This is done following the design choice already in place: when the
encoding changes, emit a signal from MousepadFile, forwarded by
MousepadDocument to MousepadWindow, which updates the status bar. When
needed, MousepadWindow asks MousepadDocument to resend all necessary
signals to update the status bar, including the encoding signal.

See !69 for more details.