Improve `--quit` option: close all windows interactively
The change to the option itself is very simple, but particular attention
is paid to the fact that triggering the option when one or more dialogs
are open causes all windows to close cleanly.

A menu entry is also added to the "File" menu, assigned `Ctrl+Q`, and
reassigning `Ctrl+Shift+W` to "Close Window".

See !71 for more details.