Properly initialize document search properties
After several patches related to the workaround to prevent endless
searches (see !58 and the commits that followed), what
`mousepad_document_search_widget_visible()` do is a kind of binding,
that need a connection (already in place) and an initialization
(almost absent). This commit properly initializes this binding.

That way, in particular, when opening a new document while a search
widget is used, the "highlight" and "regex-enabled" properties are
properly initialized, which is not currently the case.