Ensure that the page setup is properly applied when printing
For some reason (most likely a bug on GTK side, see
https://gitlab.gnome.org/GNOME/gtk/-/issues/1752), the page setup is
lost between the page setup dialog and `begin_print()`. As a workaround,
this commit saves it as a MousepadPrint attribute and apply it by hand
in `begin_print()`.

As a result, user settings regarding page setup are now properly saved
when the print operation is completed.

Fixes #90.