Block the right signal handler for tooltip updates
This is a mistake in !50 that has been inconsequential so far, because
in the current state of the code, it can probably only have a (minor)
effect on performance.

The correct signal handler to block is
`mousepad_window_menu_update_tooltips()` and not
`mousepad_window_menu_set_tooltips`, which is not connected to any
signal. Furthermore, when the blockage is released, the signal handler
must be called once manually.

This also fixes a small mistake in previous commit regarding action
groups initialization.