Use old style menu alignment
This restores the alignment of menu items in Mousepad 0.4.2, before the
deprecated GtkUIManager/GtkAction APIs were replaced (see !25). As this
has a slight impact on startup performance, and to leave a fallback
solution in case of problems, a hidden parameter is added that allows
to switch back to the native alignment (set to old-style alignment by
default).

Closes #97, see !77 for more details.