Transfer encoding from remote to primary instance
This is particular to options taking an optional argument, which are
parsed into a `GOptionArgFunc()` and do not appear in the options
dictionary.

It is therefore possible to add the parsed option to the dictionary in
`handle_local_options()`, after processing by the `GOptionArgFunc()` but
still in the remote instance, and before `command_line()`, where it can
be retrieved by the primary instance.

Fixes #109.