Delay request to scroll to cursor for the correct active view
The active view can change between the time the request to scroll to
cursor is sent and the time it is executed. While this is unlikely in
the only use case of `mousepad_window_scroll_to_cursor()` at the moment,
it could be more of a problem in the future (see
https://gitlab.xfce.org/apps/mousepad/-/merge_requests/83#note_29251).

Also, since the extension of `mousepad_view_scroll_to_cursor()` from
`void (*)(MousepadView *)` to `gboolean (*)(gpointer)` is natural, it is
better to send the scroll request directly to this function for the
correct active view.