Use get/set_real_line_offset () functions for "Paste as Column"
This makes "Paste as Column" function to insert string just before
TAB char if the column of insertion falls in the middle or start of
the TAB char and insert string at the end of TAB char if and only
if column of insertion falls at the end of TAB char.

See https://gitlab.xfce.org/apps/mousepad/-/issues/114#note_29910
for more details.