Plugins: Add gspell plugin
The plugin listen to all required signals to be aware of adding/removing
views to the application, and attaches/detaches gspell objects
accordingly. When attached to a view, on "populate-popup" signal, it
extracts the gspell menu from the GTK context menu to add it to the
Mousepad context menu.

The plugin is set as non-destroyable, so that costly spell checkers are
disabled, not destroyed, when the plugin itself is disabled, and
enabling/disabling the plugin after the first time is fast.

Closes #1, obsoletes !2, see !92 for more details.