Do a silent search when changing tabs with the search bar enabled
This is a follow-up to https://gitlab.xfce.org/apps/mousepad/-/issues/83#note_21027,
in the form of a compromise. The search context is updated in the
background, indicating in particular the new occurrences count in the
search bar, but without disturbing the user in the text view: no
selection of the search result, no scrolling to cursor.

Related to !58.