Add a test script for Mousepad
The script is divided into test sections. It has no mandatory argument:
it runs by default all tests sections for the first `mousepad` binary
found in `$PATH`. The full script usage is documented by the `--help`
option.

This first commit adds:
* A 'non-gui' section, dedicated to the few command line options of
  Mousepad that don't involve any GUI, e.g. `--version`.
* A 'simple-gui' section, which launches Mousepad with its other
  command line options, and makes it quit as soon as the window appears
  on screen.

For simplicity, `xdotool` is used for the moment to monitor window
display. This will probably change in the future in favor of a more
platform-independent tool, listening directly to GTK signals.

See !99 for more details.