test-script: Add setting change test
This tests setting change from the command line (via `gsettings`) when
Mousepad is running, in various configurations: standalone prefs dialog,
empty document, one file, two files in tabs, two files in windows.

This is roughly equivalent to testing toggle actions bound to GSettings,
that is (with a few exceptions) application actions. But it is easier to
do it via `gsettings`, and it will be easier to maintain in the future.

See !99 for more details.