Enable "Save" action for read-only documents
It is redirected to the "Save As" action via a massage dialog, to both
explain the reason to the user and ask for confirmation.

The message appearing in the same dialog when closing a modified
read-only document is also slightly modified, to indicate to the user
why "save as" is proposed instead of "save".

Closes #126, see !103 for more details.