Tests: Add a test plugin to monitor application state
This plugin is disabled by default: building with `--enable-plugin-test`
is required to enable it.

The plugin listens to all required signals to monitor application state
and informs the test script via D-Bus. For now, the use of the already
existing "is-busy" GApplication property has proven to be sufficient for
this purpose.

The only modification of Mousepad core code is the replacement of
`mousepad_application{activate,open}()` calls with
`g_application{activate,open}()` calls, so that these calls trigger the
corresponding signals emission.

See !105 for more details.