Tests: Port the window actions test to the test plugin
Again, the already exposed GApplication structure is used to communicate
with the test script via D-Bus: a test action for the window actions is
added to the application, which the script can then activate via D-Bus.

Basically, what was done in the script is transported to the plugin,
with the difference that instead of sorting the actions, we can use more
advanced data structures (structure, hash table), and thus make the code
easier to read, modify and maintain.

The dependencies to xdotool and wmctrl are now removed.

See !105 for more details.