Plugin support: Properly handle plugin settings
What was done in !92 was strictly speaking only suitable for one plugin.
This could be extended to several plugins by adding a `.gschema.xml`
file in `plugins/`, intermediate between the one of the application and
those of the plugins, or by refactoring the Makefiles with inclusions
and a single call to `@GSETTINGS_RULES@`.

But in any case, due to the relative rigidity of the `.gschema.xml` file
format and the internal workings of `glib-compile-schemas`, this would
only be suitable for plugins that are present at compile time, i.e.
"fake plugins".

Instead, this commit adds the plugin settings at load time, as is
natural and as the `GSettingsSchema` documentation states. To do this,
the setting store is extended to contain several roots: the application
root and the plugin roots.

For the latter, a unified naming convention is preserved, with the
prefix `org.xfce.mousepad.plugins.`, but they are in fact completely
independent of each other and independent of the application root.

Fixes #136, related to !92.