Fix wrong externally modified document in some situations
If the externally modified document is in the active tab when
the "externally-modified" signal is sent, it is not necessarily true
when the user focuses the window (i.e. when the "notify::is-active"
signal is sent), e.g. if he opened a new document from the command line
or the file manager in between.

Related to !64.