Move recent history management to `mousepad_util.c`
The recent manager is typically a resource that lies at the application
level, not at the window level. Since there is no particular reason to
add it to `mousepad_application.c`, nor to create a dedicated file at
this point, it is put in `mousepad_util.c`.

A few things are simplified in the process, and the encoding lookup
function is rewritten to add more data to the history later, such as
language and cursor position. All these data are naturally subject to
the privacy setting added in !89.

Related to !89, see !107 for more details.