Always use recent history encoding when opening a file
Instead of just using it when the first opening attempt fails, this
always uses it, unless the user has explicitly set an encoding, either
via the command line or the "Open" dialog.

As previously for the "Open" dialog and the encoding dialog, history
encoding roughly acts as default encoding for the given file here,
taking precedence over the global default encoding.

Related to !69, see !107 for more details.