Add session backup and restore
The list of tabs per window is saved at different times, in particular
when switching tab or active window. This way, an accurate picture of
the current session is kept at all times, in a format described in the
`.gschema.xml` file.

When activating "app.quit" or when closing the last window, session
backup is suspended to keep the current state as the one to be restored
on next session.

In general, session restore takes precedence over any command line
argument. In particular, a restored file is opened with encoding and
cursor position found in the recent history (or default values), even
if it is specified on the command line with or without any option.

Files specified on the command line that are not included in the list
of files to restore are opened normally, from the state in which the
application is at the end of the session restore, just as if they had
been opened while the application was already running.

Closes #21, see !108 for more details.