Add autosave support
This only concerns autosave itself, and does not include any restore
feature, which is covered by separate commits.

Any document open in a MousepadWindow is automatically saved when
modified, in a temporary file located in
`g_get_user_data_dir()/Mousepad` and named `autosave-N` where N is a
unique id, with a delay of `preferences.file.autosave-timer` seconds.
Setting this parameter to 0 disables the feature.

When a document is regularly saved by the user, the temporary file
associated with it is automatically deleted.

These automatic save and delete operations are asynchronous. The
GApplication use count is used to ensure that they have time to finish
correctly.

See !109 for more details.