Filter notebook click events also on y-coordinate
For some reason, probably internal to GTK, when the view is focused the
notebook is also focused. As a result, when the left mouse button is
held down and the right button is pressed in turn, this second event is
received by the notebook instead of the view.

To avoid unwanted tab menu displays, as well as the tab switching that
may accompany them, this commit ensures that the right click takes place
in the tab label, not only according to the x-coordinate but also
according to the y-coordinate.

Fixes #144.