Search: Ensure settings synchronization before searching
This should rarely be a problem in practice, but there is at least one
problematic case. Enabling the regex search automatically starts a
search to update the context. This search can then be launched before
the context has been informed of this change, and thus without the
workaround to prevent endless scanning being active, even though the
search will be done by regex as soon as the context receives the
information.

One could query the Mousepad settings rather than the context settings
to know the nature of the search in this case, but this is flawed and,
in general, the search should only start when Mousepad and the search
context are synchronized.

Related to !58.