Search: Prevent GTask warnings when replacing text
Since 844616a47d42794dc40dc6dafbc59caf47facd19 where the use of a
GCancellable has been introduced, depending on the state of the buffer
scan, a text replacement can cause a warning when GtkSourceView tries to
cancel the GTask.

To avoid this, we must first exit the callback of this GTask before
starting another search or replace action. This commit therefore splits
this callback in two parts, deferring the second part with a high
priority to preserve the real time behavior.

Related to !58.