Drop restrictions on URI scheme
What was implemented in !112 (and the related commits that followed)
for the trash can, actually allows to do without any restrictions on the
URI scheme. This completes the switch to GFile in !64, so that Mousepad
no longer necessarily relies on the existence of a local path to manage
a file, but can do so entirely via the GFile abstraction.

This allows in particular to open files via the network in https, or any
other URI scheme supported by GIO, which is left to impose its own
restrictions, collected via the error handling already in place.

That said, the existence of the file locally remains the normal and
preferred mode of use. Nothing is done in particular to overcome the
slowness that can occur when opening a file via the network, where
simple existence tests can take a few tenths of a second.

Consequently, without even invoking security issues, the Flatpak version
should remain delivered without network access by default.