Force encoding when reloading
By hypothesis the file has already been opened with the current
encoding, so the question of its validity no longer arises. Moreover, as
this is a direct call to `mousepad_file_open()`, if the encoding is not
forced and is only partially valid, the reloading of the file fails
without even opening the encoding dialog.

Related to !69.