search: Do not select first combo box entry when list does not change
Regression introduced in ea239c3f7e3b75cc75470576c6afbcafc157a302 and
19ce10e75b43f406783180f72f0d28a06ac2c470.

This causes a problem when replacing with an empty string, by then
selecting the first entry in the replacement list, thus preventing
replacing again with an empty string. It also selects the first entry in
the search list when searching for an empty string, which is less
annoying but still a bug.

More generally, there is no reason to re-select the first entry if the
list has not changed.

Related: !119