Keep saved state in memory to track actual modification state
This is implemented in a simple way, at the cost of a duplication of the
buffer content in memory (plus a second copy at string comparison time,
see comment in mousepad_file_buffer_changed() for more details). For a
simple graphical text editor, by nature not very suitable for editing
large files, this does not seem to be a problem. This can always be
optimized later on though, this first version seeks to be functionally
correct first.

Related: #170, !123