build: Optionally require polkit
Polkit has become required since 1696e660 since polkit.its is necessary
for xgettext to parse the `.policy` file. This check does not ensure
that polkit.its will be present in all cases, as it may be in a package
without a pkg-config file (example: it is in polkitd on Ubuntu 22.04).
But it should still improve things in general.

Related: !126