Don't set buffer as unmodified before reloading if externally modified
This can lead to an inconsistent state if reloading fails for any reason
(e.g. because the file has been deleted in between), making the user
believe that his file is saved when it isn't. What's more, since
autosave relies on the modified state of the buffer, this disables it by
mistake.

There is still a check when Mousepad closes, which should prevent data
loss, and this situation is unlikely in practice, but this premature
modification of the buffer state is certainly wrong.