shortcuts-plugin: Fix shortcuts-editor include
This accidently works on distros that follows FHS because libxfce4util
has -I${includedir}/xfce4 as its Cflags and libxfce4kbd-private-3 requires
that. In FHS distros includedir is mostly the same everywhere, but this is
not the case in NixOS.

In libxfce4kbd-private-3.pc we have -I${includedir}/xfce4/libxfce4kbd-private-3,
let's adjust the xfce-shortcuts-editor.h include to follow that.

Fixes: https://gitlab.xfce.org/apps/mousepad/-/issues/295